adios: re-org adios module
