+++ opendrift/models/openoil/adios/oil.py
@@ -18,17 +18,11 @@ Interface to the ADIOS oil database.
 """
 
 import logging
-
 logger = logging.getLogger(__name__)
-import requests
+
 from typing import List
 
-ADIOS = "https://adios.orr.noaa.gov/api/oils/"
-
-# The SSL configuration of the ADIOS database does not work outside the browser it seems. Please see
-# this issue: https://github.com/NOAA-ORR-ERD/adios_oil_database/issues/2 .
-VERIFY = False
-
+from . import api
 
 class ThinOil:
     """
@@ -72,8 +66,7 @@ class ThinOil:
         """
         Fetch the full oil from ADIOS.
         """
-        return Oil.from_id(self.id)
-
+        return api.get_full_oil_from_id(self.id)
 
 class Oil(ThinOil):
     data: dict
@@ -88,12 +81,6 @@ class Oil(ThinOil):
 
         from pprint import pp
         pp(o)
-
-    @staticmethod
-    def from_id(_id) -> 'Oil':
-        logger.debug(f"Fetching full oil: {_id}")
-        o = requests.get(f"{ADIOS}/{_id}", verify=VERIFY).json()
-        return Oil(o)
 
     def __repr__(self):
         return f"[<adios.Oil> {self.id}] {self.name}"
@@ -136,50 +123,3 @@ class Oil(ThinOil):
         raise NotImplementedError
 
 
-def oils(limit=50, query='') -> List[ThinOil]:
-    """
-    Get all oils.
-
-    Args:
-
-        limit: number of oils to retrieve, <= 0 means all available.
-        query: search string (name, id, location).
-
-
-    Returns:
-
-        List of `class:ThinOil`s.
-    """
-    # The batch size seems to be maximum 205 at the moment.
-    MAX_BATCH_SZ = 205
-
-    batch = min(MAX_BATCH_SZ, limit)
-
-    oils = []
-
-    # XXX: This fails when batch size is less or unequal to `batch`.
-    while len(oils) < limit or limit <= 0:
-        p = int(len(oils) / batch)  # next page, XXX: check for off-by-one?
-        logging.debug(
-            f"Requesting list of oils from ADIOS, oils: {len(oils)} of {limit}, page: {p}"
-        )
-
-        o = requests.get(ADIOS, {
-            'dir': 'asc',
-            'limit': batch,
-            'page': p,
-            'sort': 'metadata.name',
-            'q': query
-        },
-                         verify=VERIFY).json()
-
-        oils.extend(o['data'])
-
-        if float(o['meta']['totalPages']) <= p:
-            break
-
-    limit = len(oils) if limit <= 0 else limit
-    oils = oils[:min(limit, len(oils))]
-    oils = [ThinOil.from_json(o) for o in oils]
-
-    return oils
