+++ tests/models/openoil/test_adios_api.py
@@ -1,13 +1,6 @@
 import pytest
 from opendrift.models.openoil import adios
 
-@pytest.fixture
-def aasgard():
-    oils = adios.oils(1, 'AASGARD A 2003')
-    f = oils[0].make_full()
-    assert f.name == 'AASGARD A 2003'
-    return f
-
 
 def test_get_all():
     oils = adios.oils(300)
@@ -19,10 +12,8 @@ def test_get_all():
     sids = set(ids)
     assert len(ids) == len(sids)
 
+
 def test_get_full():
     thin = adios.oils(1)[0]
     full = thin.make_full()
     print(full)
-
-def test_culled_density(aasgard):
-    print(aasgard)
